requirements.txt
customtkinter
Pillow
requests
keyboard
mss
openai-whisper
sounddevice
numpy
scipy
